+++ src/Symfony/Component/Form/Extension/Core/Type/DateType.php
@@ -362,7 +362,7 @@ class DateType extends AbstractType
         $result = [];
 
         foreach ($years as $year) {
-            $result[\PHP_INT_SIZE === 4 ? \DateTime::createFromFormat('Y e', $year.' UTC')->format('U') : gmmktime(0, 0, 0, 6, 15, $year)] = $year;
+            $result[\PHP_INT_SIZE === 4 ? \DateTimeImmutable::createFromFormat('Y e', $year.' UTC')->format('U') : gmmktime(0, 0, 0, 6, 15, $year)] = $year;
         }
 
         return $result;